+++ BE-MyCarMaster/src/main/java/softeer/bemycarmaster/api/domain/wheeldrive/dto/request/GetWheelDrivesRequest.java
@@ -18,5 +18,5 @@ public class GetWheelDrivesRequest {
 	@Schema(description = "트림 식별자", example = "1")
 	@NotNull(message = "trimId는 Null일 수 없습니다.")
 	@Min(value = 1, message = "trimId는 1 이상의 값입니다.")
-	private Integer trimId;
+	private Long trimId;
 }
